third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-dom-primitiveUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'primitiveUnits' attribute of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.getAttribute('primitiveUnits') is "userSpaceOnUse"
PASS filterElement.getAttribute('primitiveUnits') is "objectBoundingBox"
PASS successfullyParsed is true

TEST COMPLETE
